Returns an av_state for data loaded from a data.frame

This function constructs an object of class av_state based on a given data.frame.

load_dataframe(df, net_cfg, log_level = 0)

Arguments

df
a data.frame with features in columns and measurements in rows.

net_cfg
a net_cfg object providing metadata about the networks

log_level
sets the minimum level of output that should be shown (a number between 0 and 3). A lower level means more verbosity. Specify a log_level of 3 to hide messages about the data set that was loaded.

Value

This function returns the newly created av_state object.

Examples

av_state <- load_dataframe(generate_test_data(),new_net_cfg())

#> load_dataframe loaded data.frame with 3 columns:
#> [1] "id (scl)"       "tijdstip (scl)" "home (nom)"    

print(av_state)

#> 
#> file_name: fake_filename.sav 
#> real_file_name: fake_filename.sav 
#> file_type: SPSS 
#> raw_data: 5 samples with 3 features [missing: 6.67% (1)]
#> data: 
#>   multiple: 5 samples with 3 features [missing: 6.67% (1)]
#>